Major reorganization of code structure. Renaming of several files to better reflect their usage.
