+++ app/__init__.py
@@ -34,10 +34,10 @@ bcrypt = Bcrypt(app)
 login_manager = LoginManager(app)
 login_manager.login_view = 'login'  # route that login_required redirects to
 
-from app.models import User, AuthToken, Athlete, Workout
+from app.db_models import User, AuthToken, Athlete, Workout
 from app import routes  # position important to avoid circular importation
 from app import admin
-Workout.__table__.drop(db.engine) # Use if you want to drop the table and reset it
+# Workout.__table__.drop(db.engine) # Use if you want to drop the table and reset it
 # Athlete.__table__.drop(db.engine) # Use if you want to drop the table and reset it
 db.create_all()  # Only creates tables when they do not already exist
 db.session.commit()
