+++ app/forms/forms.py
@@ -2,7 +2,7 @@ from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
 from wtforms.widgets import TextArea
-from app.models import User
+from app.db_models import User
 
 class RegistrationForm(FlaskForm):
     username = StringField('Username',
